+++ lib/models/dom_element.ts
@@ -1,4 +1,5 @@
 import { JSHandle, ElementHandle } from 'puppeteer';
+import { isXPathQuery } from '../utils/utils';
 
 export default class DomElement {
     public readonly element: ElementHandle;
@@ -10,25 +11,41 @@ export default class DomElement {
     }
 
     public async query(selector: string): Promise<DomElement | null> {
-        const element = await this.element.$(selector);
-        return DomElement.processQueryResult(element, selector);
-    }
-
-    public async queryXPath(selector: string): Promise<Array<DomElement>> {
-        return this.element.$x(selector).then((elements) => {
-            return elements.map((e) => {
-                return new DomElement(e, selector);
-            });
-        });
+        let elementHandle: ElementHandle | null;
+        if (isXPathQuery(selector)) {
+            const results = await this.element.$x(selector);
+            elementHandle = results[0] || null;
+        } else elementHandle = await this.element.$(selector);
+        return DomElement.processQueryResult(elementHandle, selector);
     }
 
     public async queryAll(selector: string): Promise<Array<DomElement>> {
-        return this.element.$$(selector).then((elements) => {
-            return elements.map((e) => {
-                return DomElement.processQueryResult(e, selector);
-            }).filter(b => Boolean(b)) as Array<DomElement>;
-        });
+        let elements: Array<ElementHandle>;
+        if (isXPathQuery(selector)) {
+            if (selector[0] === "/") selector = `.${selector}`;
+            elements = await this.element.$x(selector);
+        } else elements = await this.element.$$(selector);
+
+        return elements.map((e) => {
+            return DomElement.processQueryResult(e, selector);
+        }).filter(b => Boolean(b)) as Array<DomElement>;
     }
+
+    // public async queryXPath(selector: string): Promise<Array<DomElement>> {
+    //     return this.element.$x(selector).then((elements) => {
+    //         return elements.map((e) => {
+    //             return new DomElement(e, selector);
+    //         });
+    //     });
+    // }
+    //
+    // public async queryAll(selector: string): Promise<Array<DomElement>> {
+    //     return this.element.$$(selector).then((elements) => {
+    //         return elements.map((e) => {
+    //             return DomElement.processQueryResult(e, selector);
+    //         }).filter(b => Boolean(b)) as Array<DomElement>;
+    //     });
+    // }
 
     public click(): Promise<void> {
         return this.element.click();
@@ -36,6 +53,10 @@ export default class DomElement {
 
     public tap(): Promise<void> {
         return this.element.tap();
+    }
+
+    public focus(): Promise<void> {
+        return this.element.focus();
     }
 
     public toString(): string {