+++ lib/modules/webworkers/browser_webworker.ts
@@ -1,4 +1,4 @@
-import WebWorker from '../../models/webworker';
+import WebWorker from './webworker';
 import WendigoModule from '../wendigo_module';
 
 export default class BrowserWebWorker extends WendigoModule {